week-06-tasks/requirements.txt
flask
fastapi
uvicorn
pydantic
python-dotenv